+++ Hazel/Platform/OpenGL/OpenGlIndexBuffer.h
@@ -6,17 +6,17 @@ namespace Hazel {
 
    class OpenGLIndexBuffer : public IndexBuffer {
    public:
-      OpenGLIndexBuffer(const unsigned int* indices, const size_t count);
+      OpenGLIndexBuffer(const uint32_t* indices, const uint32_t count);
       virtual ~OpenGLIndexBuffer();
 
       virtual void Bind() const override;
       virtual void Unbind() const override;
 
-      virtual size_t GetCount() const override;
+      virtual uint32_t GetCount() const override;
 
    private:
-      unsigned int m_indexBufferId;
-      size_t m_count;
+      uint32_t m_indexBufferId;
+      uint32_t m_count;
    };
 
 }
