+++ Hazel/src/Renderer/IndexBuffer.cpp
@@ -4,7 +4,7 @@
 #include "Platform/OpenGL/OpenGlIndexBuffer.h"
 
 namespace Hazel {
-   IndexBuffer* IndexBuffer::Create(const unsigned int* indices, const size_t count) {
+   IndexBuffer* IndexBuffer::Create(const uint32_t* indices, const uint32_t count) {
       if (Renderer::GetRendererAPI() == RendererAPI::OpenGL) {
          return new OpenGLIndexBuffer(indices, count);
       }
